refactor: TwitterSessionPayload
